refactors cli to disco
