+++ Makefile
@@ -77,7 +77,7 @@ build-cli: tidy ## Builds CLI binary
 	-w -s -X main.commit=$(COMMIT) \
 	-w -s -X main.date=$(CURRENT_DATE) \
 	-extldflags '-static'" \
-    -a -mod vendor -o bin/disco cmd/cli/main.go
+    -a -mod vendor -o bin/disco cmd/disco/main.go
 
 .PHONY: server
 server: ## Runs previsouly built server binary
