+++ cmd/disco/cli/img_test.go
@@ -1,4 +1,4 @@
-package disco
+package cli
 
 import (
 	"flag"
