+++ cmd/disco/cli/vuln.go
@@ -1,4 +1,4 @@
-package disco
+package cli
 
 import (
 	"github.com/mchmarny/disco/pkg/disco"
@@ -8,35 +8,37 @@ import (
 )
 
 var (
-	pkgCmd = &c.Command{
-		Name:    "package",
-		Aliases: []string{"pkg"},
-		Usage:   "Scan images for packages",
-		Action:  runPackageCmd,
+	vulnCmd = &c.Command{
+		Name:    "vulnerability",
+		Aliases: []string{"vul"},
+		Usage:   "Check for OS and package level exposures in images (supports CVE filter)",
+		Action:  runVulnsCmd,
 		Flags: []c.Flag{
 			imageListPathFlag,
 			imageURIFlag,
 			outputPathFlag,
 			outputFormatFlag,
 			projectIDFlag,
+			minSeverityFlag,
+			cveFlag,
 			targetFlag,
-			pkgNameFlag,
 		},
 	}
 )
 
-func runPackageCmd(c *c.Context) error {
-	in := &types.PackageQuery{}
+func runVulnsCmd(c *c.Context) error {
+	in := &types.VulnsQuery{}
 	in.Quiet = isQuiet(c)
 	in.Version = getVersionFromContext(c)
-	in.Kind = types.KindPackage
+	in.Kind = types.KindVulnerability
 	in.ProjectID = c.String(projectIDFlag.Name)
 	in.OutputPath = c.String(outputPathFlag.Name)
+	in.CVE = c.String(cveFlag.Name)
 	in.OutputFmt = types.ParseOutputFormatOrDefault(c.String(outputFormatFlag.Name))
 	in.ImageFile = c.String(imageListPathFlag.Name)
 	in.ImageURI = c.String(imageURIFlag.Name)
+	in.MinVulnSev = types.ParseMinVulnSeverityOrDefault(c.String(minSeverityFlag.Name))
 	in.TargetRaw = c.String(targetFlag.Name)
-	in.NamePart = c.String(pkgNameFlag.Name)
 
 	printVersion(c)
 
@@ -49,8 +51,8 @@ func runPackageCmd(c *c.Context) error {
 		return errors.Wrap(err, "invalid target")
 	}
 
-	if err := disco.DiscoverPackages(c.Context, in, ir); err != nil {
-		return errors.Wrap(err, "error discovering packages")
+	if err := disco.DiscoverVulns(c.Context, in, ir); err != nil {
+		return errors.Wrap(err, "error excuting command")
 	}
 
 	return nil
